Update Gitignore, Install Cloudflare Wrangler

+++ Sources/Jobs/CleanOutJob.swift
@@ -3,7 +3,9 @@ import Foundation
 struct CleanOutJob: Job {
     var title: String { "Clean Output Folder" }
     var handler: (JobParams) throws -> Void = { params in
-        try FileManager.default.removeItem(at: params.outDir)
-        try JobHelpers.createOutDir(params.outDir)
+        if FileManager.default.fileExists(atPath: params.outDir.path()) {
+            try FileManager.default.removeItem(at: params.outDir)
+            try JobHelpers.createOutDir(params.outDir)
+        }
     }
 }
